% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phats_fn.R
\name{est_phats}
\alias{est_phats}
\title{Estimate the gene type proportions.}
\usage{
est_phats(Y, markers, baseline_ests, gamma)
}
\arguments{
\item{Y}{The expression matrix. Data frame or matrix. Each row contains expression measurements for a particular sample. Each columm contains the measurements of the same gene over all individuals.}

\item{markers}{Marker gene indices. List of vectors. List should be same length as \code{pure_samples}, i.e. one element for each cell type. Each element of the top-level list is a vector of indicies (columns of Y) that will be considered markers of that particular type. If not supplied to dtangle then markers will be determined internally by dtangle using \code{\link{find_markers}}.}

\item{baseline_ests}{List of vectors (same structure as markers). One list entry for each cell type. Each list element is a vector of estimated offset for each marker of the respective type (output from \code{baseline_exprs}).}

\item{gamma}{Expression sensitivity parameter. If provided as a single positive number then that value will be used for gamma and over-ride the ``data_type`` option.}
}
\value{
Estimated matrix of mixing proportions.
}
\description{
Estimate the gene type proportions.
}
