% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamma.R
\name{get_gamma}
\alias{get_gamma}
\title{Determine gamma value by data type.}
\usage{
get_gamma(data_type)
}
\arguments{
\item{data_type}{A string indicating the type of the expression measurements. Used to choose pre-estimated gamma value. Current support for probe-level microarray as ``microarray-probe'', gene-level microarray as ``microarray-gene'' or rna-seq as ``rna-seq''.}
}
\description{
Determine gamma value by data type.
}
